refactor(checker): extract mode-aware cross-file export resolver to own file
Move `resolve_cross_file_export_from_file_with_mode` out of
`type_resolution/module.rs` into a new `type_resolution/cross_file_export.rs`
so `module.rs` stays under its architecture size ratchet (was 1977, limit
1953; now 1825). The default-mode wrappers stay in `module.rs` and delegate.

Widens three helpers (`resolve_ambient_module_export`,
`resolve_module_augmentation_export_for_file`,
`module_augmentation_export_preempts_reexport_alias`) to `pub(crate)` so the
relocated method can reach them. Behavior-neutral.
